requirements.txt
streamlit==1.13.0
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
protobuf==3.20.0
numpy==1.21.5
pandas==1.3.5
google-api-core==2.7.1
google-auth==2.6.2
google-auth-oauthlib==0.5.1
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.3
pandas-gbq==0.17.9